refactor: split ui.go into commands, keyhandlers, and helpers
Splits the 1283-line ui.go into focused files:
- ui.go (core model, ~360 lines)
- commands.go (data fetching, ~375 lines)
- keyhandlers.go (key handling, ~365 lines)
- helpers.go (utilities, ~213 lines)

No behavior changes.
