+++ internal/tui/ui.go
@@ -2,8 +2,6 @@ package tui
 
 import (
 	"fmt"
-	"os/exec"
-	"strings"
 	"time"
 
 	"github.com/charmbracelet/bubbles/key"
@@ -359,925 +357,3 @@ func (m Model) View() string {
 	return result
 }
 
-// handleKeyPress processes keyboard input
-func (m Model) handleKeyPress(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	// When help overlay is visible, only ? and esc close it; ignore everything else
-	if m.help.Visible() {
-		if msg.String() == "?" || msg.Type == tea.KeyEscape {
-			m.help.Toggle()
-		}
-		return m, nil
-	}
-
-	// ? toggles help overlay in any mode
-	if msg.String() == "?" {
-		m.help.Toggle()
-		return m, nil
-	}
-
-	// Global quit key
-	if msg.String() == "q" || msg.String() == "ctrl+c" {
-		return m, tea.Quit
-	}
-
-	switch m.viewMode {
-	case ViewModeList:
-		return m.handleListKeys(msg)
-	case ViewModeDetail:
-		return m.handleDetailKeys(msg)
-	case ViewModeLog:
-		return m.handleLogKeys(msg)
-	case ViewModeDiff:
-		return m.handleDiffKeys(msg)
-	case ViewModeKanban:
-		return m.handleKanbanKeys(msg)
-	case ViewModeToolTimeline:
-		return m.handleToolTimelineKeys(msg)
-	case ViewModeMission:
-		return m.handleMissionKeys(msg)
-	}
-
-	return m, nil
-}
-
-// handleListKeys handles keys in list view mode
-func (m Model) handleListKeys(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "j", "down":
-		m.taskList.MoveCursor(1)
-	case "k", "up":
-		m.taskList.MoveCursor(-1)
-	case "enter":
-		// If cursor is on a collapsed group header, expand it instead of opening detail
-		if m.taskList.IsCursorOnCollapsedGroup() {
-			m.taskList.ToggleGroupExpand()
-			return m, nil
-		}
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			if session.Source == data.SourceLocalCopilot {
-				m.ctx.Error = nil
-				m.viewMode = ViewModeDetail
-				m.taskDetail.SetTask(session)
-				return m, nil
-			}
-			m.viewMode = ViewModeDetail
-			return m, m.fetchTaskDetail(session.ID, session.Repository)
-		}
-	case "l":
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			m.viewMode = ViewModeLog
-			if isSessionRunning(session) {
-				m.logView.SetLive(true)
-				m.logView.SetFollowMode(true)
-				return m, tea.Batch(m.fetchTaskLog(session.ID, session.Repository), m.logPollTick())
-			}
-			return m, m.fetchTaskLog(session.ID, session.Repository)
-		}
-case "c":
-session := m.taskList.SelectedTask()
-if session != nil && session.Source == data.SourceLocalCopilot && session.HasLog {
-m.viewMode = ViewModeLog
-m.showConversation = true
-return m, m.fetchConversation(session.ID)
-}
-	case "o":
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			return m, m.openTaskPR(session)
-		}
-	case "s":
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			return m, m.resumeSession(session)
-		}
-	case "x":
-		m.taskList.DismissSelected()
-	case "r":
-		return m, m.fetchTasks
-	case "p":
-		m.showPreview = !m.showPreview
-		m.updateSplitLayout()
-		return m, nil
-	case "a":
-		m.ctx.StatusFilter = "attention"
-		m.showPreview = false
-		m.taskList.SetLoading(true)
-		return m, m.fetchTasks
-	case "tab":
-		m.cycleFilter(1)
-		m.taskList.SetLoading(true)
-		return m, m.fetchTasks
-	case "shift+tab", "backtab":
-		m.cycleFilter(-1)
-		m.taskList.SetLoading(true)
-		return m, m.fetchTasks
-	case "g":
-		m.taskList.CycleGroupBy()
-		return m, nil
-	case " ":
-		if m.taskList.IsGrouped() {
-			m.taskList.ToggleGroupExpand()
-			return m, nil
-		}
-	case "K":
-		m.viewMode = ViewModeKanban
-		m.kanban.SetSize(m.ctx.Width, m.ctx.Height-4)
-		return m, nil
-	case "M":
-		m.viewMode = ViewModeMission
-		m.mission.SetSize(m.ctx.Width, m.ctx.Height-4)
-		return m, nil
-	case "!":
-		return m, m.openSourceRepo()
-	case "@":
-		return m, m.openFileIssue()
-	case "d":
-		session := m.taskList.SelectedTask()
-		if session != nil && canShowDiff(session) {
-			m.diffView.SetLoading()
-			m.viewMode = ViewModeDiff
-			return m, m.fetchPRDiff(session)
-		} else if session != nil {
-			m.toast.Push("⚠️", session.Title, "no PR — session is on "+session.Branch)
-		}
-	case "t":
-		session := m.taskList.SelectedTask()
-		if session != nil && session.Source == data.SourceLocalCopilot && session.HasLog {
-			m.viewMode = ViewModeToolTimeline
-			m.toolTimeline.SetSize(m.ctx.Width-4, m.ctx.Height-8)
-			return m, m.fetchToolTimeline(session.ID)
-		}
-	}
-	return m, nil
-}
-
-// handleDetailKeys handles keys in detail view mode
-func (m Model) handleDetailKeys(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.viewMode = ViewModeList
-	case "l":
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			m.viewMode = ViewModeLog
-			if isSessionRunning(session) {
-				m.logView.SetLive(true)
-				m.logView.SetFollowMode(true)
-				return m, tea.Batch(m.fetchTaskLog(session.ID, session.Repository), m.logPollTick())
-			}
-			return m, m.fetchTaskLog(session.ID, session.Repository)
-		}
-case "c":
-session := m.taskList.SelectedTask()
-if session != nil && session.Source == data.SourceLocalCopilot && session.HasLog {
-m.viewMode = ViewModeLog
-m.showConversation = true
-return m, m.fetchConversation(session.ID)
-}
-	case "o":
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			return m, m.openTaskPR(session)
-		}
-	case "s":
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			return m, m.resumeSession(session)
-		}
-	case "t":
-		session := m.taskList.SelectedTask()
-		if session != nil && session.Source == data.SourceLocalCopilot && session.HasLog {
-			m.viewMode = ViewModeToolTimeline
-			m.toolTimeline.SetSize(m.ctx.Width-4, m.ctx.Height-8)
-			return m, m.fetchToolTimeline(session.ID)
-		}
-	case "d":
-		session := m.taskList.SelectedTask()
-		if session != nil && canShowDiff(session) {
-			m.diffView.SetLoading()
-			m.viewMode = ViewModeDiff
-			return m, m.fetchPRDiff(session)
-		}
-	}
-	return m, nil
-}
-
-// handleDiffKeys handles keys in diff view mode
-func (m Model) handleDiffKeys(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.viewMode = ViewModeList
-		return m, nil
-	}
-
-	// Delegate to viewport for scrolling
-	var cmd tea.Cmd
-	m.diffView, cmd = m.diffView.Update(msg)
-	return m, cmd
-}
-func (m Model) handleLogKeys(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.viewMode = ViewModeList
-		m.logView.SetLive(false)
-		m.logView.SetFollowMode(false)
-		m.showConversation = false
-	case "c":
-		session := m.taskList.SelectedTask()
-		if session != nil && session.Source == data.SourceLocalCopilot {
-			if m.showConversation {
-				m.showConversation = false
-			} else {
-				return m, m.fetchConversation(session.ID)
-			}
-		}
-	case "j", "down":
-		if m.showConversation {
-			m.conversationView.LineDown()
-		} else {
-			m.logView.SetFollowMode(false)
-			m.logView.LineDown()
-		}
-	case "k", "up":
-		if m.showConversation {
-			m.conversationView.LineUp()
-		} else {
-			m.logView.SetFollowMode(false)
-			m.logView.LineUp()
-		}
-	case "d":
-		if m.showConversation {
-			m.conversationView.HalfPageDown()
-		} else {
-			m.logView.SetFollowMode(false)
-			m.logView.HalfPageDown()
-		}
-	case "u":
-		if m.showConversation {
-			m.conversationView.HalfPageUp()
-		} else {
-			m.logView.SetFollowMode(false)
-			m.logView.HalfPageUp()
-		}
-	case "g":
-		if m.showConversation {
-			m.conversationView.GotoTop()
-		} else {
-			m.logView.SetFollowMode(false)
-			m.logView.GotoTop()
-		}
-	case "G":
-		if m.showConversation {
-			m.conversationView.GotoBottom()
-		} else {
-			m.logView.SetFollowMode(true)
-			m.logView.GotoBottom()
-		}
-	case "f":
-		if !m.showConversation {
-			m.logView.SetFollowMode(!m.logView.FollowMode())
-			if m.logView.FollowMode() {
-				m.logView.GotoBottom()
-			}
-		}
-	case "s":
-		session := m.taskList.SelectedTask()
-		if session != nil {
-			return m, m.resumeSession(session)
-		}
-	}
-	return m, nil
-}
-
-// handleKanbanKeys handles keys in kanban view mode
-func (m Model) handleKanbanKeys(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc", "K":
-		m.viewMode = ViewModeList
-	case "h", "left":
-		m.kanban.MoveColumn(-1)
-	case "l", "right":
-		m.kanban.MoveColumn(1)
-	case "j", "down":
-		m.kanban.MoveRow(1)
-	case "k", "up":
-		m.kanban.MoveRow(-1)
-	case "enter":
-		session := m.kanban.SelectedSession()
-		if session != nil {
-			if session.Source == data.SourceLocalCopilot {
-				m.ctx.Error = nil
-				m.viewMode = ViewModeDetail
-				m.taskDetail.SetTask(session)
-				return m, nil
-			}
-			m.viewMode = ViewModeDetail
-			return m, m.fetchTaskDetail(session.ID, session.Repository)
-		}
-	case "r":
-		return m, m.fetchTasks
-	}
-	return m, nil
-}
-
-// handleToolTimelineKeys handles keys in tool timeline view mode
-func (m Model) handleToolTimelineKeys(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.viewMode = ViewModeDetail
-	case "j", "down":
-		m.toolTimeline.LineDown()
-	case "k", "up":
-		m.toolTimeline.LineUp()
-	case "d":
-		m.toolTimeline.HalfPageDown()
-	case "u":
-		m.toolTimeline.HalfPageUp()
-	}
-	return m, nil
-}
-
-// handleMissionKeys handles keys in mission control view mode
-func (m Model) handleMissionKeys(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc", "M":
-		m.viewMode = ViewModeList
-	case "j", "down":
-		m.mission.MoveCursor(1)
-	case "k", "up":
-		m.mission.MoveCursor(-1)
-	case "enter":
-		// Return to list view (enter feels natural to "drill in")
-		m.viewMode = ViewModeList
-	case "r":
-		return m, m.fetchTasks
-	}
-	return m, nil
-}
-
-// cycleFilter cycles through status filters by delta (+1 forward, -1 backward)
-func (m *Model) cycleFilter(delta int) {
-	filters := []string{"active", "completed", "failed", "all", "attention"}
-	for i, f := range filters {
-		if f == m.ctx.StatusFilter {
-			next := (i + delta) % len(filters)
-			if next < 0 {
-				next += len(filters)
-			}
-			m.ctx.StatusFilter = filters[next]
-			m.showPreview = false
-			break
-		}
-	}
-}
-
-// Message types
-type tasksLoadedMsg struct {
-	tasks       []data.Session
-	allSessions []data.Session // unfiltered, for kanban/mission
-	counts      FilterCounts
-}
-
-type taskDetailLoadedMsg struct {
-	task *data.Session
-}
-
-type taskLogLoadedMsg struct {
-	log string
-}
-
-type toolTimelineLoadedMsg struct {
-	events []tooltimeline.ToolEvent
-}
-
-type diffLoadedMsg struct {
-	files []diffview.FileDiff
-}
-
-type refreshTickMsg struct{}
-
-type animationTickMsg struct{}
-
-type logPollTickMsg struct{}
-
-type logPollResultMsg struct {
-	log string
-	err error
-}
-
-type errMsg struct {
-	err error
-}
-
-type conversationLoadedMsg struct {
-	messages []conversation.ChatMessage
-}
-
-// fetchTasks fetches the list of sessions (both agent tasks and local sessions)
-func (m Model) fetchTasks() tea.Msg {
-	sessions, err := data.FetchAllSessions(m.repo)
-	if err != nil {
-		return errMsg{err}
-	}
-
-	// Exclude dismissed sessions before computing anything
-	dismissedIDs := map[string]struct{}{}
-	if m.dismissedStore != nil {
-		dismissedIDs = m.dismissedStore.IDs()
-	}
-	visible := make([]data.Session, 0, len(sessions))
-	for _, session := range sessions {
-		if _, dismissed := dismissedIDs[session.ID]; !dismissed {
-			visible = append(visible, session)
-		}
-	}
-	sessions = visible
-
-	// Enrich sessions with token usage from CLI logs
-	tokenUsage, _ := data.FetchTokenUsage()
-	for i := range sessions {
-		if usage, ok := tokenUsage[sessions[i].ID]; ok {
-			if sessions[i].Telemetry == nil {
-				sessions[i].Telemetry = &data.SessionTelemetry{}
-			}
-			sessions[i].Telemetry.Model = usage.Model
-			sessions[i].Telemetry.InputTokens = usage.InputTokens
-			sessions[i].Telemetry.OutputTokens = usage.OutputTokens
-			sessions[i].Telemetry.CachedTokens = usage.CachedTokens
-			sessions[i].Telemetry.ModelCalls = usage.Calls
-		}
-	}
-
-	// Compute counts across all visible (non-dismissed) sessions
-	counts := FilterCounts{All: len(sessions)}
-	for _, session := range sessions {
-		if data.SessionNeedsAttention(session) {
-			counts.Attention++
-		}
-		if data.StatusIsActive(session.Status) || strings.EqualFold(session.Status, "needs-input") {
-			counts.Active++
-		}
-		if strings.EqualFold(session.Status, "completed") {
-			counts.Completed++
-		}
-		if strings.EqualFold(session.Status, "failed") {
-			counts.Failed++
-		}
-	}
-
-	// Filter sessions based on status filter
-	if m.ctx.StatusFilter != "all" {
-		filtered := []data.Session{}
-		for _, session := range sessions {
-			if m.ctx.StatusFilter == "attention" && data.SessionNeedsAttention(session) {
-				filtered = append(filtered, session)
-			} else if m.ctx.StatusFilter == "active" && (data.StatusIsActive(session.Status) || strings.EqualFold(session.Status, "needs-input")) {
-				filtered = append(filtered, session)
-			} else if strings.EqualFold(session.Status, m.ctx.StatusFilter) {
-				filtered = append(filtered, session)
-			}
-		}
-		sessions = filtered
-	}
-
-	return tasksLoadedMsg{sessions, visible, counts}
-}
-
-// fetchTaskDetail fetches detailed information for a session
-func (m Model) fetchTaskDetail(id string, repo string) tea.Cmd {
-	return func() tea.Msg {
-		// For now, we only support detail view for agent-task sessions
-		// Local sessions don't have a detail API yet
-		task, err := data.FetchAgentTaskDetail(id, repo)
-		if err != nil {
-			return errMsg{err}
-		}
-		session := data.FromAgentTask(*task)
-		return taskDetailLoadedMsg{&session}
-	}
-}
-
-// fetchTaskLog fetches the log for a task
-func (m Model) fetchTaskLog(id string, repo string) tea.Cmd {
-	session := m.taskList.SelectedTask()
-	return func() tea.Msg {
-		// Route to local session log reader for local-copilot sessions
-		if session != nil && session.Source == data.SourceLocalCopilot {
-			log, err := data.FetchLocalSessionLog(id)
-			if err != nil {
-				return errMsg{err}
-			}
-			return taskLogLoadedMsg{log}
-		}
-
-		log, err := data.FetchAgentTaskLog(id, repo)
-		if err != nil {
-			return errMsg{err}
-		}
-		return taskLogLoadedMsg{log}
-	}
-}
-
-// fetchToolTimeline fetches tool execution events for the timeline view
-func (m Model) fetchToolTimeline(sessionID string) tea.Cmd {
-	return func() tea.Msg {
-		events, err := data.FetchSessionEvents(sessionID)
-		if err != nil {
-			return errMsg{err}
-		}
-
-		var toolEvents []tooltimeline.ToolEvent
-		for _, ev := range events {
-			if ev.Type == "tool.execution_start" && ev.ToolName != "" {
-				toolEvents = append(toolEvents, tooltimeline.ToolEvent{
-					Timestamp: ev.Timestamp,
-					ToolName:  ev.ToolName,
-					Icon:      tooltimeline.ToolIcon(ev.ToolName),
-				})
-			}
-		}
-
-		return toolTimelineLoadedMsg{events: toolEvents}
-	}
-}
-
-// fetchConversation loads events for a local session and converts them to chat messages.
-func (m Model) fetchConversation(sessionID string) tea.Cmd {
-	return func() tea.Msg {
-		events, err := data.FetchSessionEvents(sessionID)
-		if err != nil {
-			return errMsg{err}
-		}
-
-		var messages []conversation.ChatMessage
-		var pendingTools []string
-
-		for _, ev := range events {
-		switch ev.Type {
-			case "session.start":
-				messages = append(messages, conversation.ChatMessage{
-					Role:      conversation.RoleSystem,
-					Content:   "Session started",
-					Timestamp: ev.Timestamp,
-				})
-			case "user.message":
-				if ev.Content != "" {
-					messages = append(messages, conversation.ChatMessage{
-						Role:      conversation.RoleUser,
-						Content:   ev.Content,
-						Timestamp: ev.Timestamp,
-					})
-				}
-			case "tool.execution_start":
-				if ev.ToolName != "" {
-					pendingTools = append(pendingTools, ev.ToolName)
-				}
-			case "assistant.message":
-				if ev.Content != "" {
-					messages = append(messages, conversation.ChatMessage{
-						Role:      conversation.RoleAssistant,
-						Content:   ev.Content,
-						Timestamp: ev.Timestamp,
-						Tools:     pendingTools,
-					})
-					pendingTools = nil
-				}
-			case "abort":
-				messages = append(messages, conversation.ChatMessage{
-					Role:      conversation.RoleSystem,
-					Content:   "Session aborted",
-					Timestamp: ev.Timestamp,
-				})
-			}
-		}
-
-		return conversationLoadedMsg{messages: messages}
-	}
-}
-
-func (m Model) openSourceRepo() tea.Cmd {
-	return func() tea.Msg {
-		_, err := data.RunGH("repo", "view", "maxbeizer/gh-agent-viz", "--web")
-		if err != nil {
-			return errMsg{fmt.Errorf("failed to open repo: %w", err)}
-		}
-		return nil
-	}
-}
-
-func (m Model) openFileIssue() tea.Cmd {
-	return func() tea.Msg {
-		_, err := data.RunGH("issue", "create", "-R", "maxbeizer/gh-agent-viz", "--web")
-		if err != nil {
-			return errMsg{fmt.Errorf("failed to open issue form: %w", err)}
-		}
-		return nil
-	}
-}
-
-func (m Model) openTaskPR(session *data.Session) tea.Cmd {
-	return func() tea.Msg {
-		if session == nil {
-			return errMsg{fmt.Errorf("no session selected")}
-		}
-
-		// If we already have PR info, use it
-		if session.PRURL != "" {
-			output, err := data.RunGH("pr", "view", session.PRURL, "--web")
-			if err != nil {
-				return errMsg{fmt.Errorf("failed to open PR: %s", strings.TrimSpace(string(output)))}
-			}
-			return nil
-		}
-		if session.PRNumber > 0 && session.Repository != "" {
-			output, err := data.RunGH("pr", "view", fmt.Sprintf("%d", session.PRNumber), "-R", session.Repository, "--web")
-			if err != nil {
-				return errMsg{fmt.Errorf("failed to open PR: %s", strings.TrimSpace(string(output)))}
-			}
-			return nil
-		}
-
-		// Try to discover PR by branch name
-		if session.Repository != "" && session.Branch != "" {
-			prNumber, prURL, _ := data.FetchPRForBranch(session.Repository, session.Branch)
-			if prURL != "" {
-				session.PRNumber = prNumber
-				session.PRURL = prURL
-				output, err := data.RunGH("pr", "view", prURL, "--web")
-				if err != nil {
-					return errMsg{fmt.Errorf("failed to open PR: %s", strings.TrimSpace(string(output)))}
-				}
-				return nil
-			}
-		}
-
-		return errMsg{fmt.Errorf("no PR found for this branch")}
-	}
-}
-
-// resumeSessionErr returns an error tea.Cmd if the session cannot be resumed,
-// or nil if the session is valid for resumption.
-func resumeSessionErr(session *data.Session) tea.Cmd {
-	if session == nil {
-		return func() tea.Msg { return errMsg{fmt.Errorf("no session selected")} }
-	}
-	if session.Source != data.SourceLocalCopilot {
-		return func() tea.Msg { return errMsg{fmt.Errorf("only local Copilot CLI sessions can be resumed")} }
-	}
-	normalizedStatus := strings.ToLower(strings.TrimSpace(session.Status))
-	if normalizedStatus != "running" && normalizedStatus != "queued" && normalizedStatus != "needs-input" {
-		return func() tea.Msg {
-			return errMsg{fmt.Errorf("cannot resume: session status is '%s' — only running, queued, or needs-input sessions are resumable", session.Status)}
-		}
-	}
-	if session.ID == "" {
-		return func() tea.Msg { return errMsg{fmt.Errorf("cannot resume session: session has no ID")} }
-	}
-	return nil
-}
-
-func (m Model) resumeSession(session *data.Session) tea.Cmd {
-	if errCmd := resumeSessionErr(session); errCmd != nil {
-		return errCmd
-	}
-
-	// Use tea.ExecProcess to hand terminal control to the interactive
-	// Copilot CLI resume command. This suspends the TUI, lets the child
-	// process use stdin/stdout directly, and resumes the TUI on exit.
-	c := exec.Command("gh", "copilot", "--", "--resume", session.ID)
-	return tea.ExecProcess(c, func(err error) tea.Msg {
-		if err != nil {
-			return errMsg{fmt.Errorf("resume session exited with error: %w", err)}
-		}
-		return nil
-	})
-}
-
-func (m Model) logPollTick() tea.Cmd {
-	return tea.Tick(2*time.Second, func(t time.Time) tea.Msg {
-		return logPollTickMsg{}
-	})
-}
-
-func (m Model) fetchLogPoll(id string, repo string, source data.SessionSource) tea.Cmd {
-	return func() tea.Msg {
-		var log string
-		var err error
-		if source == data.SourceLocalCopilot {
-			log, err = data.FetchLocalSessionLog(id)
-		} else {
-			log, err = data.FetchAgentTaskLog(id, repo)
-		}
-		return logPollResultMsg{log: log, err: err}
-	}
-}
-
-func isSessionRunning(session *data.Session) bool {
-	return session != nil && data.StatusIsActive(session.Status)
-}
-
-func (m Model) refreshCmd() tea.Cmd {
-	return tea.Tick(m.refreshInt, func(time.Time) tea.Msg {
-		return refreshTickMsg{}
-	})
-}
-
-func (m Model) animationTickCmd() tea.Cmd {
-	return tea.Tick(100*time.Millisecond, func(time.Time) tea.Msg {
-		return animationTickMsg{}
-	})
-}
-
-func isValidFilter(filter string) bool {
-	switch filter {
-	case "all", "attention", "active", "completed", "failed":
-		return true
-	default:
-		return false
-	}
-}
-
-// smartDefaultFilter picks the best starting tab based on actual session counts.
-func smartDefaultFilter(counts FilterCounts) string {
-	if counts.Active > 0 {
-		return "active"
-	}
-	if counts.Attention > 0 {
-		return "attention"
-	}
-	return "all"
-}
-
-// previewVisible returns true when the split-pane detail preview should render.
-func (m Model) previewVisible() bool {
-	return m.showPreview && m.ctx.Width >= 80 && m.ctx.Height > 20
-}
-
-// updateSplitLayout recalculates component dimensions for the current layout.
-func (m *Model) updateSplitLayout() {
-	if m.previewVisible() {
-		leftWidth := m.ctx.Width * 2 / 5
-		rightWidth := m.ctx.Width - leftWidth
-		contentHeight := m.ctx.Height - 4 // header + footer chrome
-		m.taskList.SetSize(leftWidth, contentHeight)
-		m.taskList.SetSplitMode(true)
-		m.taskDetail.SetSize(rightWidth, contentHeight)
-	} else {
-		m.taskList.SetSize(m.ctx.Width, m.ctx.Height)
-		m.taskList.SetSplitMode(false)
-	}
-	m.kanban.SetSize(m.ctx.Width, m.ctx.Height-4)
-}
-
-// updateFooterHints updates footer hints based on current view mode and state
-func (m *Model) updateFooterHints() {
-	switch m.viewMode {
-	case ViewModeList:
-		hints := []key.Binding{
-			key.NewBinding(key.WithKeys("↑/↓"), key.WithHelp("↑/↓", "navigate")),
-			m.keys.SelectTask,
-			m.keys.ToggleFilter,
-			m.keys.ToggleKanban,
-			m.keys.ToggleMission,
-			m.keys.ShowHelp,
-			m.keys.ExitApp,
-		}
-		m.footer.SetHints(hints)
-	case ViewModeDetail:
-		hints := []key.Binding{
-			m.keys.NavigateBack,
-			m.keys.ShowLogs,
-		}
-		session := m.taskList.SelectedTask()
-		if session != nil && session.Source == data.SourceLocalCopilot && session.HasLog {
-			hints = append(hints, key.NewBinding(key.WithKeys("t"), key.WithHelp("t", "tools")))
-		}
-		// Show diff hint when session has a PR
-		if canShowDiff(session) {
-			hints = append(hints, m.keys.ShowDiff)
-		}
-		hints = append(hints, m.keys.ShowHelp, m.keys.ExitApp)
-		m.footer.SetHints(hints)
-	case ViewModeLog:
-		logHints := []key.Binding{
-			m.keys.NavigateBack,
-			key.NewBinding(key.WithKeys("↑/↓"), key.WithHelp("↑/↓", "scroll")),
-			m.keys.ToggleFollow,
-		}
-		session := m.taskList.SelectedTask()
-		if session != nil && session.Source == data.SourceLocalCopilot {
-			logHints = append(logHints, key.NewBinding(key.WithKeys("c"), key.WithHelp("c", "convo")))
-		}
-		logHints = append(logHints, m.keys.ShowHelp, m.keys.ExitApp)
-		m.footer.SetHints(logHints)
-	case ViewModeKanban:
-		kanbanHints := []key.Binding{
-			m.keys.NavigateBack,
-			key.NewBinding(key.WithKeys("h/l"), key.WithHelp("h/l", "column")),
-			key.NewBinding(key.WithKeys("j/k"), key.WithHelp("j/k", "card")),
-			m.keys.ShowHelp,
-			m.keys.ExitApp,
-		}
-		m.footer.SetHints(kanbanHints)
-	case ViewModeToolTimeline:
-		timelineHints := []key.Binding{
-			m.keys.NavigateBack,
-			key.NewBinding(key.WithKeys("↑/↓"), key.WithHelp("↑/↓", "scroll")),
-			m.keys.ShowHelp,
-			m.keys.ExitApp,
-		}
-		m.footer.SetHints(timelineHints)
-	case ViewModeMission:
-		missionHints := []key.Binding{
-			m.keys.NavigateBack,
-			key.NewBinding(key.WithKeys("j/k"), key.WithHelp("j/k", "navigate")),
-			m.keys.SelectTask,
-			m.keys.ShowHelp,
-			m.keys.ExitApp,
-		}
-		m.footer.SetHints(missionHints)
-	case ViewModeDiff:
-		diffHints := []key.Binding{
-			m.keys.NavigateBack,
-			key.NewBinding(key.WithKeys("↑/↓"), key.WithHelp("↑/↓", "scroll")),
-			m.keys.ShowHelp,
-			m.keys.ExitApp,
-		}
-		m.footer.SetHints(diffHints)
-	}
-}
-
-func canShowLogs(session *data.Session) bool {
-	if session == nil || strings.TrimSpace(session.ID) == "" {
-		return false
-	}
-	if session.Source == data.SourceAgentTask {
-		return true
-	}
-	// Local sessions can show logs if they have an events.jsonl file
-	return session.HasLog
-}
-
-func canOpenPR(session *data.Session) bool {
-	if session == nil {
-		return false
-	}
-	// Has explicit PR info
-	if strings.TrimSpace(session.PRURL) != "" {
-		return true
-	}
-	if session.PRNumber > 0 && strings.TrimSpace(session.Repository) != "" {
-		return true
-	}
-	// Can discover PR by branch lookup
-	return strings.TrimSpace(session.Repository) != "" && strings.TrimSpace(session.Branch) != ""
-}
-
-func canResumeLocalSession(session *data.Session) bool {
-	if session == nil || session.Source != data.SourceLocalCopilot || strings.TrimSpace(session.ID) == "" {
-		return false
-	}
-	status := strings.ToLower(strings.TrimSpace(session.Status))
-	return status == "running" || status == "queued" || status == "needs-input"
-}
-
-// canShowDiff returns true when the session has a PR or can discover one
-func canShowDiff(session *data.Session) bool {
-	if session == nil {
-		return false
-	}
-	if session.PRNumber > 0 && strings.TrimSpace(session.Repository) != "" {
-		return true
-	}
-	// Can discover PR by branch
-	return strings.TrimSpace(session.Repository) != "" && strings.TrimSpace(session.Branch) != ""
-}
-
-// fetchPRDiff fetches the PR diff for a session, discovering the PR by branch if needed
-func (m Model) fetchPRDiff(session *data.Session) tea.Cmd {
-	prNumber := session.PRNumber
-	repo := session.Repository
-	branch := session.Branch
-	return func() tea.Msg {
-		// Try to discover PR if not already known
-		if prNumber == 0 && repo != "" && branch != "" {
-			n, url, _ := data.FetchPRForBranch(repo, branch)
-			if n > 0 {
-				prNumber = n
-				session.PRNumber = n
-				session.PRURL = url
-			}
-		}
-		if prNumber == 0 {
-			return errMsg{fmt.Errorf("no PR found for this branch")}
-		}
-		raw, err := data.FetchPRDiff(prNumber, repo)
-		if err != nil {
-			return errMsg{err}
-		}
-		files := diffview.ParseUnifiedDiff(raw)
-		return diffLoadedMsg{files}
-	}
-}
